Revert "Move CloudFormation to separate package, expand and introduce tests"
This reverts commit 8f3dfa9076c5f9f2c531035e4e40726e9c08a83b.
